backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-modeltranslation==0.18.11
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
faker==20.1.0
slack-sdk==3.26.1
facebook-business==19.0.0
drf-yasg==1.21.7
sentry-sdk==1.39.2